This doesn't work.
Further updates. Now the connections are split into two sides of the
program to allow a peer-to-peer-type interaction. It's probably not the
best way, but PyMessage only implements a simple client that may
initiate connections and sends and receives in sequence. PyMessageServer
implements a client that may ACCEPT connections, then receives and sends
in sequence. Right now, neither one connects to the other correctly and
even if they did they wouldn't work right.