No more Python.
Scrapped the file trying to accomplish the messenger project in python.
Currently, I'm using a name that somebody else is almost definitely
using and I'll change that by the Java version of the program is up. And
it works fairly well. Next item on the task list is to develop some kind
of encoding or encryption system.